pyproject.toml
[project]
name = "gcp-iam-janitor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-auth>=2.41.1",
    "google-auth-httplib2>=0.2.0",
    "google-cloud-iam>=2.19.1",
    "google-cloud-resource-manager>=1.14.2",
    "networkx>=3.5",
    "numpy>=2.3.3",
    "pandas>=2.3.3",
    "plotly>=6.3.1",
    "python-dotenv>=1.1.1",
    "pyvis>=0.3.2",
    "streamlit>=1.50.0",
    "streamlit-aggrid>=1.1.9",
]
